+++ kcidb/orm/__init__.py
@@ -1588,7 +1588,7 @@ def argparse_add_args(parser):
     )
 
 
-class ArgumentParser(kcidb.misc.ArgumentParser):
+class ArgumentParser(kcidb.argparse.ArgumentParser):
     """
     Command-line argument parser with common ORM arguments added.
     """
@@ -1605,7 +1605,7 @@ class ArgumentParser(kcidb.misc.ArgumentParser):
         argparse_add_args(self)
 
 
-class OutputArgumentParser(kcidb.misc.OutputArgumentParser):
+class OutputArgumentParser(kcidb.argparse.OutputArgumentParser):
     """
     Command-line argument parser for tools outputting JSON,
     with common ORM arguments added.
